Move ProfileContext to contexts/*
